refactor: extract accounts into accounts-rds.ts
